refactor: split wind and temperature content into dedicated files
